separarate CostOfTheRoute logic
