+++ src/Services/Resources/CostOfTheRouteFilter.cs
@@ -1,0 +1,11 @@
+using Models;
+
+namespace Services.Resources
+{
+    public struct CostOfTheRouteFilter
+    {
+        public Node[] nodes;
+        public bool cheapestCostOfTheRoute;
+        public Graph graph;
+    }
+}
